+++ python/paddle/incubate/complex/helper.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from ..fluid import framework
+from ...fluid import framework
 
 
 def is_complex(x):
