+++ python/paddle/incubate/complex/tensor/manipulation.py
@@ -14,8 +14,8 @@
 
 from paddle.common_ops_import import *
 from ..helper import is_complex, is_real, complex_variable_exists
-from ...fluid.framework import ComplexVariable
-from ...fluid import layers
+from ....fluid.framework import ComplexVariable
+from ....fluid import layers
 
 __all__ = [
     'reshape',
