+++ python/paddle/incubate/complex/tensor/math.py
@@ -14,9 +14,9 @@
 
 from paddle.common_ops_import import *
 from ..helper import is_complex, is_real, complex_variable_exists
-from ...fluid.framework import ComplexVariable
-from ...fluid import layers
-from ...tensor import math
+from ....fluid.framework import ComplexVariable
+from ....fluid import layers
+from ....tensor import math
 
 __all__ = [
     'elementwise_add',
