+++ python/setup.py.in
@@ -139,8 +139,8 @@ packages=['paddle',
           'paddle.dataset',
           'paddle.reader',
           'paddle.distributed',
-          'paddle.complex',
-          'paddle.complex.tensor',
+          'paddle.incubate.complex',
+          'paddle.incubate.complex.tensor',
           'paddle.framework',
           'paddle.fluid',
           'paddle.fluid.dygraph',
